Move some solar eclipse computations from AstroCalcDialog
This reduces the size of AstroCalcDialog.cpp by 15% while
localizing much of solar eclipse map generation and some other
related computations in another file.
