+++ po/stellarium/POTFILES.in
@@ -26,6 +26,7 @@ src/core/modules/MilkyWay.cpp
 src/core/modules/ZodiacalLight.cpp
 src/core/modules/ToastMgr.cpp
 src/core/modules/HipsMgr.cpp
+src/core/SolarEclipseComputer.cpp
 src/core/modules/SpecificTimeMgr.cpp
 src/core/StelApp.cpp
 src/core/StelObserver.cpp
